@@ -174,7 +174,7 @@ function getPrDescriptionMarkdown(submitterNameMarkdown, fixtureWarningsErrorsMa
   let body = changedFiles.map(line => `* ${line}`).join(`\n`);
 
   if (githubErrors.length > 0) {
-    const githubErrorList = githubErrors.map(error => `* :warning: ${error}`).join(`\n`);
+    const githubErrorList = githubErrors.map(error => `* ⚠️ ${error}`).join(`\n`);
     body += `\n\n### Errors:\n${githubErrorList}`;
   }
 
@@ -208,8 +208,8 @@ function getFixtureWarningsErrorsMarkdownList(fixtures, errors, warnings) {
     const fixtureWarnings = warnings[fixtureKey] || [];
 
     const messages = [
-      ...fixtureErrors.map(error => `  - :x: ${error}`),
-      ...fixtureWarnings.map(warning => `  - :warning: ${warning}`),
+      ...fixtureErrors.map(error => `  - ❌ ${error}`),
+      ...fixtureWarnings.map(warning => `  - ⚠️ ${warning}`),
     ].join(`\n`);
 
     if (messages.length > 0) {
