@@ -35,7 +35,7 @@ export default [
      * @returns {boolean} true if the fixture uses the feature
      */
     hasFeature: fixture => fixture.coarseChannels.some(
-      channel => channel.hasCapabilities && channel.fineChannelAliases.length > 0,
+      channel => channel.fineChannelAliases.length > 0 && channel.capabilities.length > 1,
     ),
   },
 ];
