@@ -1,5 +1,6 @@
 import childProcess from 'child_process';
 import brokenLinkChecker from 'broken-link-checker';
+import terminate from 'terminate/promise.js';
 
 import importJson from './import-json.js';
 
@@ -35,7 +36,7 @@ export default class SiteCrawler extends EventTarget {
 
       // wait until server has started
       this.serverProcess.stdout.on(`data`, output => {
-        if (output.includes(`Listening on: ${BASE_URL}`)) {
+        if (output.includes(`Listening on:`)) {
           resolve();
         }
       });
@@ -129,11 +130,7 @@ export default class SiteCrawler extends EventTarget {
    * @returns {Promise<object>} Promise that resolves with stdout and stderr as soon as the server process is closed.
    */
   async stopServer() {
-    const onServerClosed = new Promise((resolve, reject) => {
-      this.serverProcess.on(`exit`, resolve);
-    });
-    this.serverProcess.kill();
-    await onServerClosed;
+    await terminate(this.serverProcess.pid);
 
     return { stdout: this.serverStdout, stderr: this.serverStderr };
   }
