@@ -37,7 +37,16 @@ const segmentsPerFile = 4;
  * @returns {Promise<object[], Error>} The generated files.
  */
 export async function exportFixtures(fixtures, options) {
-  return fixtures.flatMap(fixture => fixture.modes.flatMap(mode => getFilesForMode(mode)));
+  return fixtures.flatMap(fixture => fixture.modes.flatMap(mode => {
+    try {
+      return getFilesForMode(mode);
+    }
+    catch (error) {
+      throw new Error(`Exporting fixture mode ${fixture.manufacturer.key}/${fixture.key}/${mode.shortName} failed: ${error}`, {
+        cause: error,
+      });
+    }
+  }));
 }
 
 /**
