@@ -8,7 +8,7 @@ import SwitchingChannel from '../../lib/model/SwitchingChannel.js';
 import ddf3FunctionGroups from './ddf3-function-groups.js';
 import ddf3Functions from './ddf3-functions.js';
 
-export const version = `0.1.1`;
+export const version = `0.1.2`;
 
 /**
  * @param {Fixture[]} fixtures An array of Fixture objects.
@@ -24,35 +24,55 @@ export async function exportFixtures(fixtures, options) {
   for (const fixture of fixtures) {
     // add device for each mode
     for (const mode of fixture.modes) {
-      const xml = xmlbuilder.begin()
-        .declaration(`1.0`, `utf-8`)
-        .element({
-          device: {
-            '@type': `DMXDevice`,
-            '@dmxaddresscount': mode.channelKeys.length,
-            '@dmxcversion': 3,
-            '@ddfversion': options.displayedPluginVersion ?? version,
-          },
+      try {
+        deviceDefinitions.push(exportFixtureMode(fixture, mode, options));
+      }
+      catch (error) {
+        throw new Error(`Exporting fixture mode ${fixture.manufacturer.key}/${fixture.key}/${mode.shortName} failed: ${error}`, {
+          cause: error,
         });
-
-      addInformation(xml, mode);
-      addFunctions(xml, mode);
-      addProcedures(xml, mode);
-
-      deviceDefinitions.push({
-        name: sanitize(`${fixture.manufacturer.key}-${fixture.key}-${(mode.shortName)}.xml`).replaceAll(/\s+/g, `-`),
-        content: xml.end({
-          pretty: true,
-          indent: `  `,
-        }),
-        mimetype: `application/xml`,
-        fixtures: [fixture],
-        mode: mode.shortName,
-      });
+      }
     }
   }
 
   return deviceDefinitions;
+}
+
+/**
+ * @param {Fixture} fixture The fixture to export.
+ * @param {Mode} mode The mode to export.
+ * @param {object} options Global options.
+ * @param {string} options.baseDirectory Absolute path to OFL's root directory.
+ * @param {Date} options.date The current time.
+ * @param {string | undefined} options.displayedPluginVersion Replacement for plugin version if the plugin version is used in export.
+ * @returns {object} The generated file.
+ */
+function exportFixtureMode(fixture, mode, options) {
+  const xml = xmlbuilder.begin()
+    .declaration(`1.0`, `utf-8`)
+    .element({
+      device: {
+        '@type': `DMXDevice`,
+        '@dmxaddresscount': mode.channelKeys.length,
+        '@dmxcversion': 3,
+        '@ddfversion': options.displayedPluginVersion ?? version,
+      },
+    });
+
+  addInformation(xml, mode);
+  addFunctions(xml, mode);
+  addProcedures(xml, mode);
+
+  return {
+    name: sanitize(`${fixture.manufacturer.key}-${fixture.key}-${(mode.shortName)}.xml`).replaceAll(/\s+/g, `-`),
+    content: xml.end({
+      pretty: true,
+      indent: `  `,
+    }),
+    mimetype: `application/xml`,
+    fixtures: [fixture],
+    mode: mode.shortName,
+  };
 }
 
 /**
