@@ -23,20 +23,14 @@ export async function exportFixtures(fixtures, options) {
   const files = fixtures.map(fixture => {
     usedManufacturers.add(fixture.manufacturer.key);
 
-
-    const jsonData = JSON.parse(JSON.stringify(fixture.jsonObject));
-    jsonData.$schema = `https://raw.githubusercontent.com/OpenLightingProject/open-fixture-library/schema-${displayedPluginVersion}/schemas/fixture.json`;
-
-    jsonData.fixtureKey = fixture.key;
-    jsonData.manufacturerKey = fixture.manufacturer.key;
-    jsonData.oflURL = fixture.url;
-
-    return {
-      name: `${fixture.manufacturer.key}/${fixture.key}.json`,
-      content: fixtureJsonStringify(jsonData),
-      mimetype: `application/ofl-fixture`,
-      fixtures: [fixture],
-    };
+    try {
+      return getFixtureFile(fixture, displayedPluginVersion);
+    }
+    catch (error) {
+      throw new Error(`Exporting fixture ${fixture.manufacturer.key}/${fixture.key} failed: ${error}`, {
+        cause: error,
+      });
+    }
   });
 
   const manufacturers = await importJson(`../../fixtures/manufacturers.json`, import.meta.url);
@@ -57,4 +51,25 @@ export async function exportFixtures(fixtures, options) {
   });
 
   return files;
+}
+
+/**
+ * @param {Fixture} fixture The fixture to export.
+ * @param {string} displayedPluginVersion The plugin version that should be displayed in the exported file.
+ * @returns {object} The generated fixture JSON file.
+ */
+function getFixtureFile(fixture, displayedPluginVersion) {
+  const jsonData = structuredClone(fixture.jsonObject);
+  jsonData.$schema = `https://raw.githubusercontent.com/OpenLightingProject/open-fixture-library/schema-${displayedPluginVersion}/schemas/fixture.json`;
+
+  jsonData.fixtureKey = fixture.key;
+  jsonData.manufacturerKey = fixture.manufacturer.key;
+  jsonData.oflURL = fixture.url;
+
+  return {
+    name: `${fixture.manufacturer.key}/${fixture.key}.json`,
+    content: fixtureJsonStringify(jsonData),
+    mimetype: `application/ofl-fixture`,
+    fixtures: [fixture],
+  };
 }
