@@ -8,8 +8,8 @@
         :schema-property="schemaProperty.items"
         :required="required || dimensionsSpecified"
         :hint="hints[0]"
-        @focus.native="onFocus()"
-        @blur.native="onBlur($event)" />
+        @focus="onFocus()"
+        @blur="onBlur($event)" />
     </Validate>
     &times;
     <Validate :state="formstate" tag="span">
@@ -19,8 +19,8 @@
         :schema-property="schemaProperty.items"
         :required="required || dimensionsSpecified"
         :hint="hints[1]"
-        @focus.native="onFocus()"
-        @blur.native="onBlur($event)" />
+        @focus="onFocus()"
+        @blur="onBlur($event)" />
     </Validate>
     &times;
     <Validate :state="formstate" tag="span">
@@ -30,8 +30,8 @@
         :schema-property="schemaProperty.items"
         :required="required || dimensionsSpecified"
         :hint="hints[2]"
-        @focus.native="onFocus()"
-        @blur.native="onBlur($event)" />
+        @focus="onFocus()"
+        @blur="onBlur($event)" />
     </Validate>
     {{ unit }}
   </span>
@@ -46,16 +46,23 @@ export default {
     PropertyInputNumber,
   },
   model: {
-    prop: `dimensions`,
+    prop: `model-value`,
+    event: `update:model-value`,
   },
   props: {
-    dimensions: arrayProp().withDefault(null),
+    modelValue: arrayProp().withDefault(null),
     hints: arrayProp().withDefault(() => [`x`, `y`, `z`]),
     schemaProperty: objectProp().required,
     unit: stringProp().optional,
     required: booleanProp().withDefault(false),
     name: stringProp().required,
     formstate: objectProp().required,
+  },
+  emits: {
+    'update:model-value': dimensions => true,
+    focus: () => true,
+    blur: () => true,
+    'vf:validate': validationData => true,
   },
   data() {
     return {
@@ -67,30 +74,30 @@ export default {
   computed: {
     x: {
       get() {
-        return this.dimensions ? this.dimensions[0] : null;
+        return this.modelValue ? this.modelValue[0] : null;
       },
       set(xInput) {
-        this.$emit(`input`, getDimensionsArray(xInput, this.y, this.z));
+        this.$emit(`update:model-value`, getDimensionsArray(xInput, this.y, this.z));
       },
     },
     y: {
       get() {
-        return this.dimensions ? this.dimensions[1] : null;
+        return this.modelValue ? this.modelValue[1] : null;
       },
       set(yInput) {
-        this.$emit(`input`, getDimensionsArray(this.x, yInput, this.z));
+        this.$emit(`update:model-value`, getDimensionsArray(this.x, yInput, this.z));
       },
     },
     z: {
       get() {
-        return this.dimensions ? this.dimensions[2] : null;
+        return this.modelValue ? this.modelValue[2] : null;
       },
       set(zInput) {
-        this.$emit(`input`, getDimensionsArray(this.x, this.y, zInput));
+        this.$emit(`update:model-value`, getDimensionsArray(this.x, this.y, zInput));
       },
     },
     dimensionsSpecified() {
-      return this.dimensions !== null;
+      return this.modelValue !== null;
     },
   },
   mounted() {
