@@ -11,8 +11,8 @@
         :required="required || rangeIncomplete"
         :hint="startHint"
         lazy
-        @focus.native="onFocus($event)"
-        @blur.native="onBlur($event)" />
+        @focus="onFocus($event)"
+        @blur="onBlur($event)" />
     </Validate>
     …
     <Validate :state="formstate" tag="span">
@@ -25,8 +25,8 @@
         :required="required || rangeIncomplete"
         :hint="endHint"
         lazy
-        @focus.native="onFocus($event)"
-        @blur.native="onBlur($event)" />
+        @focus="onFocus($event)"
+        @blur="onBlur($event)" />
     </Validate>
     {{ unit }}
   </span>
@@ -41,10 +41,11 @@ export default {
     PropertyInputNumber,
   },
   model: {
-    prop: `range`,
+    prop: `model-value`,
+    event: `update:model-value`,
   },
   props: {
-    range: arrayProp().withDefault(null),
+    modelValue: arrayProp().withDefault(null),
     name: stringProp().required,
     startHint: stringProp().withDefault(`start`),
     endHint: stringProp().withDefault(`end`),
@@ -54,6 +55,14 @@ export default {
     unit: stringProp().optional,
     required: booleanProp().withDefault(false),
     formstate: objectProp().required,
+  },
+  emits: {
+    'update:model-value': range => true,
+    'start-updated': () => true,
+    'end-updated': () => true,
+    focus: () => true,
+    blur: () => true,
+    'vf:validate': validationData => true,
   },
   data() {
     return {
@@ -66,24 +75,24 @@ export default {
   computed: {
     start: {
       get() {
-        return this.range ? this.range[0] : null;
+        return this.modelValue ? this.modelValue[0] : null;
       },
       set(startInput) {
-        this.$emit(`input`, getRange(startInput, this.end));
+        this.$emit(`update:model-value`, getRange(startInput, this.end));
         this.$emit(`start-updated`);
       },
     },
     end: {
       get() {
-        return this.range ? this.range[1] : null;
+        return this.modelValue ? this.modelValue[1] : null;
       },
       set(endInput) {
-        this.$emit(`input`, getRange(this.start, endInput));
+        this.$emit(`update:model-value`, getRange(this.start, endInput));
         this.$emit(`end-updated`);
       },
     },
     rangeIncomplete() {
-      return this.range && (this.start === null || this.end === null);
+      return this.modelValue && (this.start === null || this.end === null);
     },
   },
   mounted() {
