@@ -41,7 +41,7 @@
         :schema-property="physicalProperties.DMXconnector"
         addition-hint="other DMX connector" />
       <Validate
-        v-if="physical.DMXconnector === `[add-value]`"
+        v-if="modelValue.DMXconnector === `[add-value]`"
         :state="formstate"
         tag="span">
         <PropertyInputText
@@ -111,12 +111,11 @@
 <script>
 import { objectProp, stringProp } from 'vue-ts-types';
 import {
-  schemaDefinitions,
-  physicalProperties,
   physicalBulbProperties,
   physicalLensProperties,
+  physicalProperties,
+  schemaDefinitions,
 } from '../../../lib/schema-properties.js';
-import { clone } from '../../assets/scripts/editor-utils.js';
 
 import LabeledInput from '../LabeledInput.vue';
 import PropertyInputDimensions from '../PropertyInputDimensions.vue';
@@ -135,12 +134,16 @@ export default {
     PropertyInputText,
   },
   model: {
-    prop: `physical`,
+    prop: `model-value`,
+    event: `update:model-value`,
   },
   props: {
-    physical: objectProp().required,
+    modelValue: objectProp().required,
     formstate: objectProp().required,
     namePrefix: stringProp().required,
+  },
+  emits: {
+    'update:model-value': value => true,
   },
   data() {
     return {
@@ -148,17 +151,17 @@ export default {
       physicalProperties,
       physicalBulbProperties,
       physicalLensProperties,
-      localPhysical: clone(this.physical),
+      localPhysical: structuredClone(this.modelValue),
     };
   },
   watch: {
     localPhysical: {
       handler() {
-        this.$emit(`input`, clone(this.localPhysical));
+        this.$emit(`update:model-value`, structuredClone(this.localPhysical));
       },
       deep: true,
     },
-    'physical.DMXconnector': async function(newValue) {
+    'modelValue.DMXconnector': async function(newValue) {
       if (newValue === `[add-value]` && this.$root._oflRestoreComplete) {
         await this.$nextTick();
         this.$refs.newDmxConnectorInput.focus();
