@@ -206,7 +206,6 @@
 import { stringProp } from 'vue-ts-types';
 import Fixture from '../../../lib/model/Fixture.js';
 import Manufacturer from '../../../lib/model/Manufacturer.js';
-import { clone } from '../../assets/scripts/editor-utils.js';
 
 import A11yDialog from '../A11yDialog.vue';
 import DownloadButton from '../DownloadButton.vue';
@@ -238,6 +237,10 @@ export default {
     endpoint: stringProp().required,
     githubUsername: stringProp().optional,
     githubComment: stringProp().optional,
+  },
+  emits: {
+    success: () => true,
+    reset: () => true,
   },
   data() {
     return {
@@ -348,7 +351,7 @@ export default {
     async validate(requestBody) {
       this.requestBody = requestBody;
 
-      console.log(`validate`, clone(this.requestBody));
+      console.log(`validate`, structuredClone(this.requestBody));
 
       this.state = `validating`;
       try {
@@ -383,7 +386,7 @@ export default {
         githubComment: this.githubComment ?? null,
       };
 
-      console.log(`submit`, clone(this.requestBody));
+      console.log(`submit`, structuredClone(this.requestBody));
 
       this.state = `uploading`;
       try {
