@@ -1,6 +1,6 @@
-<template functional>
-  <time :datetime="props.date.toISOString()" :title="props.date.toISOString()">{{
-    props.date.toISOString().replace(/T.*?$/, ``)
+<template>
+  <time :datetime="isoDate" :title="isoDate">{{
+    isoDate.replace(/T.*?$/, ``)
   }}</time>
 </template>
 
@@ -10,6 +10,11 @@ import { instanceOfProp } from 'vue-ts-types';
 export default {
   props: {
     date: instanceOfProp(Date).required,
+  },
+  computed: {
+    isoDate() {
+      return this.date.toISOString();
+    },
   },
 };
 </script>
